backend/requirements.txt
# LangGraph & LangChain (2026-02 최신)
langgraph>=1.0.8
langgraph-checkpoint-postgres>=2.0.0
langchain>=1.2.10
langchain-core>=1.2.11
langchain-openai>=1.1.9

# MCP
mcp>=0.9.0

# API 서버
fastapi>=0.109.0
uvicorn>=0.27.0

# 데이터베이스
psycopg[binary,pool]>=3.2.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# 벡터 데이터베이스
qdrant-client>=1.7.0

# 유틸
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
